refactor(cli): split pydantic importer facade
